Neo4jTestAgent/genai-integration-langchain/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
fsspec==2024.12.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.65
langchain-neo4j==0.4.0
langchain-openai==0.3.22
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
neo4j==5.28.1
neo4j-graphrag==1.7.0
numpy==2.3.0
openai==1.86.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pip==25.1.1
pydantic==2.11.5
pydantic_core==2.33.2
pypdf==5.6.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
xxhash==3.5.0
zstandard==0.23.0